Added Health widget

+++ Source/TBSGame/ActorComponents/Health/HealthComponent.cpp
@@ -1,6 +1,5 @@
 // Copyright Crespoter Interactive
 
-
 #include "ActorComponents/Health/HealthComponent.h"
 
 UHealthComponent::UHealthComponent()
@@ -9,19 +8,19 @@ UHealthComponent::UHealthComponent()
 	CurrentHealth = MaxHealth;
 }
 
-uint8 UHealthComponent::GetCurrentHealth() const
+int32 UHealthComponent::GetCurrentHealth() const
 {
 	return CurrentHealth;
 }
 
-uint8 UHealthComponent::GetMaxHealth() const
+int32 UHealthComponent::GetMaxHealth() const
 {
 	return MaxHealth;
 }
 
-void UHealthComponent::Damage(const uint8 DamageAmount)
+void UHealthComponent::Damage(const int32 DamageAmount)
 {
-	const uint8 OldHealth = CurrentHealth;
+	const int16 OldHealth = CurrentHealth;
 	CurrentHealth = FMath::Clamp(CurrentHealth - DamageAmount, 0, MaxHealth);
 	// Notify even if the health values are same. Let the dependents decide
 	// how to handle a clamped health value.
@@ -33,9 +32,9 @@ void UHealthComponent::Damage(const uint8 DamageAmount)
 	}
 }
 
-void UHealthComponent::Heal(const uint8 HealAmount)
+void UHealthComponent::Heal(const int32 HealAmount)
 {
-	const uint8 OldHealth = CurrentHealth;
+	const int16 OldHealth = CurrentHealth;
 	CurrentHealth = FMath::Clamp(CurrentHealth + HealAmount, 0, MaxHealth);
 	HealthChangedDelegate.Broadcast(OldHealth, CurrentHealth);
 }
