+++ Source/TBSGame/ActorComponents/Health/HealthComponent.h
@@ -8,26 +8,26 @@
 #include "HealthComponent.generated.h"
 
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_TwoParams(FTBSHealthUpdatedDelegate,
-	uint8, OldHealth, uint8, NewHealth);
+	int32, OldHealth, int32, NewHealth);
 
 UCLASS( ClassGroup=(Custom), meta=(BlueprintSpawnableComponent) )
 class TBSGAME_API UHealthComponent : public UActorComponent
 {
 	GENERATED_BODY()
 
-public:	
+public:
 	// Sets default values for this component's properties
 	UHealthComponent();
 
 	UFUNCTION(BlueprintCallable)
-	FORCEINLINE uint8 GetCurrentHealth() const;
+	FORCEINLINE int32 GetCurrentHealth() const;
 
 	UFUNCTION(BlueprintCallable)
-	FORCEINLINE uint8 GetMaxHealth() const;
+	FORCEINLINE int32 GetMaxHealth() const;
 
-	void Damage(const uint8 DamageAmount);
+	void Damage(const int32 DamageAmount);
 
-	void Heal(const uint8 HealAmount);
+	void Heal(const int32 HealAmount);
 	
 	UPROPERTY(BlueprintAssignable)
 	FTBSMulticastDynamicDelegate HealthDepletedDelegate;
@@ -40,9 +40,9 @@ protected:
 	virtual void BeginPlay() override;
 	
 	UPROPERTY(EditAnywhere)
-	uint8 MaxHealth {10};
+	int32 MaxHealth {10};
 
 	UPROPERTY(EditAnywhere)
-	uint8 CurrentHealth {0};
+	int32 CurrentHealth {0};
 
 };
